# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mint Bucks promotional credit system for Mint Printworks
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/credits-over-time:
    get:
      operationId: getCreditsOverTime
      tags: [reports]
      summary: Credits issued and redeemed per month
      parameters:
        - name: months
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Time series data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditTimePoint"
  /reports/top-customers:
    get:
      operationId: getTopCustomers
      tags: [reports]
      summary: Top customers by credits received
      parameters:
        - name: limit
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Top customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerCreditStat"
  /reports/expiring-soon:
    get:
      operationId: getExpiringSoon
      tags: [reports]
      summary: Credits expiring within N days
      parameters:
        - name: days
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Expiring credits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Credit"

  # ── Settings ────────────────────────────────────────────────────────────────
...
  /reports/activity:
    get:
      operationId: getRecentActivity
      tags: [reports]
      summary: Recent activity feed (credits + redemptions)
      parameters:
        - name: limit
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ── Rewards ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Credit:
      type: object
      required: [id, customerId, customerName, code, amount, amountRemaining, status, issuedAt, createdAt, updatedAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerCompany:
          type: ["string", "null"]
        code:
          type: string
        amount:
          type: number
        amountRemaining:
          type: number
        status:
          type: string
          enum: [active, partially_redeemed, redeemed, expired, cancelled]
        note:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        issuedAt:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    CreditTimePoint:
      type: object
      required: [month, issued, redeemed]
      properties:
        month:
          type: string
        issued:
          type: number
        redeemed:
          type: number
    CustomerCreditStat:
      type: object
      required: [customerId, customerName, customerEmail, totalIssued, totalRedeemed, outstandingBalance, creditCount]
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerCompany:
          type: ["string", "null"]
        totalIssued:
          type: number
        totalRedeemed:
          type: number
        outstandingBalance:
          type: number
        creditCount:
          type: integer
    ActivityItem:
      type: object
      required: [id, type, customerId, customerName, amount, occurredAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [issued, redeemed, expired, cancelled]
        customerId:
          type: integer
        customerName:
          type: string
        creditCode:
          type: string
        amount:
          type: number
        invoiceRef:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        occurredAt:
          type: string
...