# lib/api-zod/src/generated/types/customerCreditStat.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mint Bucks promotional credit system for Mint Printworks
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerCreditStat {
  customerId: number;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerCompany?: string | null;
  totalIssued: number;
  totalRedeemed: number;
  outstandingBalance: number;
  creditCount: number;
}